+++ pylon_camera/include/pylon_camera/pylon_camera_node.h
@@ -938,6 +938,14 @@ protected:
      */
     bool setTriggerTimeoutCallback(camera_control_msgs::SetIntegerValue::Request &req, camera_control_msgs::SetIntegerValue::Response &res);
 
+    /**
+     * Service callback for setting the camera grabbing strategy 
+     * @param req request
+     * @param res response
+     * @return true on success
+     */
+    bool setGrabbingStrategyCallback(camera_control_msgs::SetIntegerValue::Request &req, camera_control_msgs::SetIntegerValue::Response &res);
+
     ros::NodeHandle nh_;
     PylonCameraParameter pylon_camera_parameter_set_;
     ros::ServiceServer set_binning_srv_;
@@ -986,6 +994,7 @@ protected:
     ros::ServiceServer gamma_enable_srv;
     ros::ServiceServer set_grab_timeout_srv;
     ros::ServiceServer set_trigger_timeout_srv;
+    ros::ServiceServer set_grabbing_strategy_srv;
 
     std::vector<ros::ServiceServer> set_user_output_srvs_;